correct commonly misspelled English words in source files
I recently ran across the `misspell` utility [1] and thought I'd see
how many typos it would catch in our repo. It could be neat to run this
as part of CI (maybe via the GitHub Checks API?), although we'd have to
come up with a proper blacklist of files that we'd want to ignore.

For this PR, I ran `misspell -w .` at the top-level directory of a clean
checkout (with no submodules), and then manually reverted the changes
on vendored files. The only mistake I noticed was a recommendation to
change "hardlinked" to "hardline", so I manually added the missing
space...otherwise, the suggestions seem pretty reliable. Note that
Terraform _description_ fields also had to be ignored, or unfortunately,
TF will try to replace existing resources.

[1]: https://github.com/client9/misspell